new branch test

+++ src/main/java/com/ventaproductos/orderitem/entity/OrderItemDTOUpdate.java
@@ -1,7 +1,5 @@
 package com.ventaproductos.orderitem.entity;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
@@ -13,9 +11,9 @@ public record OrderItemDTOUpdate(
     @NotNull(message =  "required field") 
     Integer orderId,
     @NotNull(message =  "required field") 
-    BigDecimal quantity,
+    int quantity,
     @NotNull(message =  "required field") 
-    BigDecimal unitPrice
+    double unitPrice
 
 ) {
